Changed to Promise-based function rather than callback.
I've done this because "returning" out of a callback is a PITA and the previous setup had order of execution issues and I didn't want to just use timeouts until async callbacks had executed.
Also moved most channel.sends to the MessageHandler because as mentioned, returning a string from the function call with async methods are very annoying. Much easier to just pass the channel as a parameter into the function itself and calling it from there.
Further housekeeping as well, syntax errors, if condition refactoring etc etc.