Shift towards google sheets
The output from a straight up DB seems like it would be a gigantic pain and also not visually engaging. Interacting from discord to a google sheet should provide a much better link for the user. Remains to be seen whether a db is even required.

More importantly, moved auth credentials into a gitignored json file.